# cuda : use amd wave sharing intrinsics for warp_reduce functions