The warp_reduce functions use `__shfl_xor_sync()` which on AMD HIP gets turned into the `ds_bpermute` instruction which requires some setup instructions and uses local data share bandwidth and latency. AMD GPUs also support Data-Parallel Primitive (DPP) instructions which can perform certain forms of sharing within a warp for free or nearly free.

Using the DPP intrinsic in this PR my GPU gets 1% faster token generation times, but oddly 0-1% slower prompt processing times. Maybe it has to do with there being 4 fewer waits on LDS memory each reduction and how the GPU schedules warps, but I really have no idea. Other GPU models might behave differently.

Details on the instructions available for sharing within a warp are on [AMD's website](https://gpuopen.com/learn/amd-gcn-assembly-cross-lane-operations/) and within each [ISA's documentation](https://gpuopen.com/amd-isa-documentation/).
I'd like to note that in the event that a warp_reduce function is called with some threads inactive, that the results will differ between GFX9 GPUs and GFX10+, but I don't think this matters since according to the [CUDA documentation](https://docs.nvidia.com/cuda/cuda-c-programming-guide/#warp-shuffle-description) performing `__shfl_xor_sync()` with inactive threads is undefined on NVIDIA GPUs anyways.

Here are my llama-bench results:
| GPU        | Model                |   Model Size [GiB] | Test   |   t/s master |   t/s PR |   Speedup |
|:-----------|:---------------------|-------------------:|:-------|-------------:|----------------------:|----------:|
| RX 5700 XT | llama 7B Q4_K_S      |               3.59 | pp512  |       300.78 |                298.79 |      0.99 |
| RX 5700 XT | llama 7B Q4_K_S      |               3.59 | tg128  |        46.63 |                 47.18 |      1.01 |
| RX 5700 XT | llama 7B Q8_0        |               6.67 | pp512  |       345.31 |                341.53 |      0.99 |
| RX 5700 XT | llama 7B Q8_0        |               6.67 | tg128  |        38.99 |                 39.38 |      1.01 |
| RX 5700 XT | mpt 7B Q4_0          |               3.64 | pp512  |       339.65 |                339.53 |      1.00 |
| RX 5700 XT | mpt 7B Q4_0          |               3.64 | tg128  |        63.27 |                 63.90 |      1.01 |
| RX 5700 XT | orion 14B Q3_K_S     |               5.96 | pp512  |        84.05 |                 83.88 |      1.00 |
| RX 5700 XT | orion 14B Q3_K_S     |               5.96 | tg128  |        15.91 |                 16.10 |      1.01 |
| RX 5700 XT | phi2 3B Q8_0         |               2.75 | pp512  |       783.94 |                783.35 |      1.00 |
| RX 5700 XT | phi2 3B Q8_0         |               2.75 | tg128  |        76.15 |                 76.67 |      1.01 |
| RX 5700 XT | starcoder2 7B Q4_K_M |               4.22 | pp512  |       280.40 |                279.28 |      1.00 |
| RX 5700 XT | starcoder2 7B Q4_K_M |               4.22 | tg128  |        39.39 |                 39.67 |      1.01 |

I also tried testing SOFT_MAX and saw between 0-8% speedup except for cases with `[1024,1024,1,1]` or `[1023,1023,1,1]` that had max bias of 8, which were ~5% slower. Again I'm not sure why.