@JohannesGaessler 2024-04-07
Sorry, what I said was misleading. What I meant to say is that I also only observe a small speedup. But I am **not** observing a performance regression. These are the specific numbers:

| GPU     | Model           |   Model Size [GiB] | Test   |   t/s master |   t/s a37d8856 |   Speedup |
|:--------|:----------------|-------------------:|:-------|-------------:|---------------:|----------:|
| RX 6800 | llama 7B Q4_0   |               3.83 | pp512  |       715.17 |         715.80 |      1.00 |
| RX 6800 | llama 7B Q4_0   |               3.83 | tg128  |        61.00 |          61.11 |      1.00 |
| RX 6800 | llama 7B Q4_K_S |               3.86 | pp512  |       586.22 |         586.72 |      1.00 |
| RX 6800 | llama 7B Q4_K_S |               3.86 | tg128  |        43.33 |          43.74 |      1.01 |
| RX 6800 | llama 7B Q8_0   |               7.17 | pp512  |       726.42 |         726.25 |      1.00 |
| RX 6800 | llama 7B Q8_0   |               7.17 | tg128  |        41.46 |          41.70 |      1.01 |