@JohannesGaessler 2024-05-10
@Engininja2 reported an inconsistent and minimal change in performance and there have been no further updates from him since. As such I don't think it makes sense to merge this PR in its current state, especially when I don't have a good understanding of what this code actually does.