@sorasoras 2024-05-11
@Engininja2 it seems like recent merge is failed compile on my 7900XTX at windows.
I got some decent performance for larger batch inference.