@Engininja2 2024-05-11
I rebased the PR on current master. llama.cpp doesn't spend much time synchronizing between GPU threads so a tiny performance increase probably isn't worth the maintenance burden of additional code.

Would this be better converted to a draft PR, or closed?