@JohannesGaessler 2024-05-11
Check performance for 
```
./llama-bench --model models/opt/${model_name}-${quantization}.gguf -r 1 -fa 1 -n 0 -pg 0,0 -p 4096 -b 1,512
```
with an actual path to a model. The new FlashAttention kernels (I think) are more sensitive to the changes in this PR.