@JohannesGaessler 2024-05-11
These are the results I get:

| GPU     | Model         |   Batch size | Test   |   t/s fed01084 |   t/s 9e6f2e2a |   Speedup |
|:--------|:--------------|-------------:|:-------|---------------:|---------------:|----------:|
| RX 6800 | llama 8B Q8_0 |            1 | pp4096 |          27.25 |          28.16 |      1.03 |
| RX 6800 | llama 8B Q8_0 |          512 | pp4096 |         380.49 |         374.30 |      0.98 |