@Engininja2 2024-05-12
Unfortunately I get a large performance regression for flash attention.
| GPU        | Model                |   Batch size | Test   |   t/s master |   t/s amd-warp-reduce |   Speedup |
|:-----------|:---------------------|-------------:|:-------|-------------:|----------------------:|----------:|
| RX 5700 XT | llama 7B Q4_K_S      |            1 | pp4096 |        23.62 |                 23.55 |      1.00 |
| RX 5700 XT | llama 7B Q4_K_S      |          512 | pp4096 |       211.17 |                163.15 |      0.77 |
| RX 5700 XT | starcoder2 7B Q4_K_M |            1 | pp4096 |        21.12 |                 21.47 |      1.02 |
| RX 5700 XT | starcoder2 7B Q4_K_M |          512 | pp4096 |       193.14 |                149.18 |      0.77 |

For gfx1010 `flash_attn_vec_ext_f16<128, 8, 1>`, the instance llama-bench is using here, spills 26 vgprs on master, and 185 on this PR. Applying `launch_bounds(D,1)` has it spill 55 vgprs, but it's using twice as many vgprs, for 177.58 t/s, 16% slower than master.

I tried replacing directly adding half2s with adding the x and y components and got this result with 75 vgprs spilled:
| model                          |    n_batch |          test |              t/s |
| ------------------------------ | ---------: | ------------: | ---------------: |
| llama 7B Q4_K - Small          |          1 |        pp4096 |     23.72 ± 0.00 |
| llama 7B Q4_K - Small          |        512 |        pp4096 |    201.41 ± 0.00 |

It's better but still a 5% slowdown, and feels like a bad idea to be hoping that compiler heuristics happen to go the right way.