@JohannesGaessler 2024-05-12
Generally speaking the current FlashAttention kernels are bad for large batch sizes anyways so the real question is how the performance will change for kernels that are specifically written for large batch sizes. I think a PR that gives +3% performance in some cases but -23% in others should not be merged.