@sorasoras 2024-05-13
It does provide decent speed up for token generation about 2-3% which is useful.