@jeroen-mostert 2024-08-29
FWIW, compared to the current master, I can't replicate the regression on my RX 6800 XT, and in fact for large batches the speedup increases accordingly when using flash attention. I did not observe any change in token generation speed (positive or negative).

| GPU        | Model         |   Batch size | FlashAttention   | Test   |   t/s master |   t/s dpp_wave |   Speedup |
|:-----------|:--------------|-------------:|:-----------------|:-------|-------------:|---------------:|----------:|
| RX 6800 XT | llama 7B Q4_0 |            1 | No               | pp512  |        79.08 |          79.52 |      1.01 |
| RX 6800 XT | llama 7B Q4_0 |            1 | No               | pp4096 |        44.76 |          44.85 |      1.00 |
| RX 6800 XT | llama 7B Q4_0 |            1 | Yes              | pp512  |        80.56 |          81.17 |      1.01 |
| RX 6800 XT | llama 7B Q4_0 |            1 | Yes              | pp4096 |        59.32 |          60.13 |      1.01 |
| RX 6800 XT | llama 7B Q4_0 |          512 | No               | pp512  |      1685.17 |        1694.99 |      1.01 |
| RX 6800 XT | llama 7B Q4_0 |          512 | No               | pp4096 |      1123.16 |        1124.75 |      1.00 |
| RX 6800 XT | llama 7B Q4_0 |          512 | Yes              | pp512  |      1553.67 |        1628.39 |      1.05 |
| RX 6800 XT | llama 7B Q4_0 |          512 | Yes              | pp4096 |       773.28 |         869.77 |      1.12 |
| RX 6800 XT | llama 7B Q4_0 |         4096 | No               | pp512  |      1684.82 |        1687.73 |      1.00 |
| RX 6800 XT | llama 7B Q4_0 |         4096 | No               | pp4096 |      1111.66 |        1112.76 |      1.00 |
| RX 6800 XT | llama 7B Q4_0 |         4096 | Yes              | pp512  |      1551.82 |        1627.11 |      1.05 |
| RX 6800 XT | llama 7B Q4_0 |         4096 | Yes              | pp4096 |       767.88 |         863.89 |      1.13 |

So this may be worth revisiting and/or gating behind specific model checks. A 12% boost is nothing to sneeze at.